Add extra braces around std::array initializers
This appears to be a bug in older clang and gcc compilers where they
gave a warning asking for an extra unnecessary set of braces around the
initializer of std::array. See for example

https://bugs.llvm.org/show_bug.cgi?id=21629
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=25137

I don't think newer compilers give this warning, but we still support
some older compilers it happens on, so we live with it.